Added frontend, still need to fix manual white balancing and camera presets
