+++ templates/index.html
@@ -1,0 +1,87 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Image Processing Tool</title>
+    <!-- Link to the CSS file using Flask's url_for -->
+    <link rel="stylesheet" href="{{ url_for('static', filename='styles.css') }}">
+    <style>
+        /* Add styles for the selection rectangle */
+        #selection-rectangle {
+            position: absolute;
+            border: 2px solid red;
+            pointer-events: none; /* Allow clicks to pass through */
+            display: none; /* Initially hidden */
+        }
+
+        /* Style for instructions */
+        #instructions {
+            margin-top: 20px;
+            font-weight: bold;
+        }
+    </style>
+</head>
+<body>
+    <div id="app">
+        <h1>Image Processing Tool</h1>
+        
+        <div id="upload-section">
+            <input type="file" id="image-upload" accept="image/*">
+            <div id="image-preview"></div>
+        </div>
+        
+        <div id="white-balance-section">
+            <h2>White Balance Method</h2>
+            <select id="white-balance-method">
+                <option value="white">White Balance</option>
+                <option value="grey">Grey Balance</option>
+               <!-- <option value="camera">Camera Presets</option> -->
+            </select>
+            <!-- <button id="manual-white-balance">Manual White Balance</button> -->
+                <button id="download-sample-image">Download Sample Image</button>
+        </div>
+        
+        
+        
+
+
+
+        
+        <div id="processing-section">
+            <button id="process-image">Process Image</button>
+            <div id="progress-bar-container" style="display: none;">
+                <div id="progress-bar"></div>
+            </div>
+        </div>
+        
+        <div id="result-section">
+            <canvas id="result-canvas"></canvas>
+            <div id="compression-options">
+                <label for="jpeg-quality">JPEG Quality:</label>
+                <input type="range" id="jpeg-quality" min="0" max="100" value="95">
+                <span id="file-size"></span>
+            </div>
+            <button id="download-image">Download Processed Image</button>
+            <button id="download-pdf-report">Download Report</button>
+
+        </div>
+        
+        <!-- Error message section -->
+        <div id="error-message">
+            <!-- Error messages will be inserted here by JavaScript -->
+        </div>
+
+       
+
+        <!-- Instructions for white patch selection -->
+        <!--- <div id="instructions">Please select a white patch on the image.</div> -->
+
+        <!-- Selection Rectangle -->
+        <div id="selection-rectangle"></div>
+    </div>
+    
+    <!-- Link to the JavaScript file using Flask's url_for -->
+    <script src="{{ url_for('static', filename='app.js') }}"></script>
+</body>
+</html>
